+++ tree_queries/forms.py
@@ -7,6 +7,8 @@ class TreeNodeIndentedLabels(object):
     def __init__(self, queryset, *args, **kwargs):
         if hasattr(queryset, "with_tree_fields"):
             queryset = queryset.with_tree_fields()
+        if "label_from_instance" in kwargs:
+            self.label_from_instance = kwargs.pop("label_from_instance")
         super(TreeNodeIndentedLabels, self).__init__(queryset, *args, **kwargs)
 
     def label_from_instance(self, obj):