+++ tree_queries/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (0, 2, 1)
+VERSION = (0, 3, 0)
 __version__ = ".".join(map(str, VERSION))